+++ apps/app/src/middleware.ts
@@ -1,9 +1,18 @@
-import { linksMiddleware, pipeMiddleware } from "@lightdotso/middlewares";
+import {
+  excludedMiddleware,
+  linksMiddleware,
+  authMiddleware,
+  pipeMiddleware,
+} from "@lightdotso/middlewares";
 import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
 
 export const middleware = (req: NextRequest) => {
-  return pipeMiddleware(req, NextResponse.next(), [linksMiddleware]);
+  return pipeMiddleware(req, NextResponse.next(), [
+    linksMiddleware,
+    excludedMiddleware,
+    authMiddleware,
+  ]);
 };
 
 export const config = {
