+++ apps/changelog/src/middleware.ts
@@ -1,9 +1,16 @@
-import { linksMiddleware, pipeMiddleware } from "@lightdotso/middlewares";
+import {
+  linksMiddleware,
+  pipeMiddleware,
+  homeMiddleware,
+} from "@lightdotso/middlewares";
 import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
 
 export const middleware = (req: NextRequest) => {
-  return pipeMiddleware(req, NextResponse.next(), [linksMiddleware]);
+  return pipeMiddleware(req, NextResponse.next(), [
+    linksMiddleware,
+    homeMiddleware,
+  ]);
 };
 
 export const config = {
