+++ packages/middlewares/src/index.ts
@@ -1,3 +1,4 @@
+export { authMiddleware } from "./auth";
 export { excludedMiddleware } from "./excluded";
 export { homeMiddleware } from "./home";
 export { linksMiddleware } from "./links";
